Add files via upload
This is a bar chart of the mature status of the streamers channels. It shows the data for if the streamers viewers label their account as mature and from the data, only a small amount of the streamers do so. The data is split between "false" and "true" and the false bar on the chart is extremely higher than the "true" bar.