Add files via upload
Before removing the outliers, the correlation between Average Viewers and Watch Time (Minutes) was 0.48, showing a moderate positive relationship. There were two really big outliers on the far right that stretched out the plot and made it harder to see the main pattern in the data.
After taking those outliers out, the new correlation went up to 0.65, which shows a moderately strong positive relationship. This means that as the average number of viewers increases, the total watch time also tends to go up. Removing those two extreme points made the scatterplot look more clear and gave a better picture of how the two variables are actually related for most streamers.